requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn>=0.24.0
pydantic>=2.0.0        
pandas>=2.0.0          
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0